+++ src/main/java/io/xlorey/FluxLoader/server/api/IncomingPacket.java
@@ -33,5 +33,4 @@ public class IncomingPacket {
         }
         return false;
     }
-
 }