+++ src/main/java/io/xlorey/FluxLoader/server/core/CommandsManager.java
@@ -73,7 +73,7 @@ public class CommandsManager {
         if (commandsMap.containsKey(commandName)) {
             throw new Exception(String.format("The '%s' command is already registered in the system!", commandName));
         }
-        Logger.print(String.format("Added new custom command: '%s'", commandName));
+        Logger.printLog(String.format("Added new custom command: '%s'", commandName));
 
         commandsMap.put(commandName, command);
     }
@@ -143,7 +143,7 @@ public class CommandsManager {
 
         String[] commandArgsToInvoke = Arrays.copyOfRange(commandArgs, 1, commandArgs.length);
         String playerName = playerConnection == null ? "Console" : playerConnection.username;
-        Logger.print(String.format("Player '%s' called command '%s' with arguments: '%s'", playerName, commandArgs[0], Arrays.toString(commandArgsToInvoke)));
+        Logger.printLog(String.format("Player '%s' called command '%s' with arguments: '%s'", playerName, commandArgs[0], Arrays.toString(commandArgsToInvoke)));
         command.onInvoke(playerConnection, commandArgsToInvoke);
 
         CommandChatReturn chatReturnAnnotation = command.getClass().getAnnotation(CommandChatReturn.class);